requirements.txt
gitpython
PyGithub
requests
